Fragmented OOF in inline crash
When an OOF is fragmented, we wait to lay it out at the fragmentation
context root. If the OOF lives above a column spanner, though, we may
lay it out early to make sure the OOF contributes to the correct column
block-size (CL:3018723).

However, if a re-attach is triggered, we may end up with old fragments
sticking around until a node has completed layout (as a result of
CL:3393649).

This created an issue when an OOF is laid out early as a result of a
columns spanner break. The OOF algorithm would attempt to access
old fragments since they are no longer cleared out by default in
LayoutBox::AddLayoutResult(). To handle this, remove all old items
not only when the node has completed layout, but also if an item broke
as a result of a column spanner.

Bug: 1290093
Change-Id: I9b47ea681534f59ad6d1ed0662238306fcc03381